Seperate Qwen3-Next and Qwen3.5 modeling functions
